requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests==2.31.0
chardet>=5.0.0